requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
Authlib==1.2.1
bardapi==0.1.38
bcrypt==4.0.1
beautifulsoup4==4.12.2
blinker==1.7.0
browser-cookie3==0.19.1
cachelib==0.10.2
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
deep-translator==1.11.4
distro==1.8.0
exceptiongroup==1.1.3
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-DebugToolbar==0.13.1
Flask-Login==0.6.3
Flask-OAuthlib==0.9.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
frozenlist==1.4.0
google-api-core==2.14.0
google-api-python-client==2.107.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-cloud-core==2.3.3
google-cloud-translate==3.12.1
googleapis-common-protos==1.61.0
greenlet==3.0.1
grpcio==1.59.2
grpcio-status==1.59.2
gunicorn==21.2.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.25.1
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
langdetect==1.0.9
lz4==4.3.2
MarkupSafe==2.1.3
multidict==6.0.4
oauth2client==4.1.3
oauthlib==2.1.0
openai==1.3.8
packaging==23.2
proto-plus==1.22.3
protobuf==4.25.0
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodomex==3.19.0
pydantic==2.5.1
pydantic_core==2.14.3
pyOpenSSL==23.3.0
pyparsing==3.1.1
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.1.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
tqdm==4.66.1
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.1.0
Werkzeug==3.0.1
WTForms==3.1.1
yarl==1.9.3
zipp==3.17.0
